requirements.txt
fastapi>=0.95.1
uvicorn>=0.22.0
pydantic>=2.0.0
supabase>=1.0.3
python-dotenv>=1.0.0
asyncpg>=0.27.0
openai>=1.0.0
pydantic-ai>=0.1.0
logfire>=3.0.0
pytest>=7.3.1
requests>=2.31.0
rich>=13.0.0
setuptools>=45.0.0